Disable allow_mass_assignment_of tests under Rails 5
The `allow_mass_assignment_of` matcher tests `attr_accessible` and
`attr_protected` which was moved to the `protected_attributes` gem in
Rails 4. The gem does not work in Rails 5 and neither does the matcher,
so there is no need to test it. (We should probably remove the matcher
altogether eventually -- Strong Parameters have long replaced
`protected_attributes`.)