Rails 5: Skip array-column-related tests for uniqueness matcher
For some reason, in Rails 5, the uniqueness validation has stopped
working for array columns. What this means is that if you try to use an
array column for a scope, then the query that the validation performs
blows up because the value in the query for the scope is not an array
but a regular scalar value. This does not seem to be documented
anywhere, so perhaps this never worked to begin with, or it accidentally
worked in Rails 4. Until we solve this mystery, skip the tests that have
to do with array columns.