Fix broken acceptance tests
Bundler 1.6.0 made a change which prohibits you from running `bundle
init` from a directory from which Bundler is able to find a Gemfile by
ascending the file tree. This breaks our acceptance tests because some
of them create a directory and then run `bundle init` in order to
bootstrap that directory as a fresh Bundler project.