Override the way gems are installed on Travis
It seems that Travis has made a change where instead of installing gems
globally they are installed to `vendor/gems`. This happens because
all of our Appraisal Gemfiles have associated lockfiles that are present
in the repo and Travis picks up on this and choose to run `bundle
with `--deployment`. This breaks our acceptance tests as they assume
that gems are installed globally.

To get around this we need to define our own `install` command.

(cherry picked from commit 309d1ff5aa5fbfbbc5bb814a67eea28b2bc14b4e)