Make it possible to run YARD locally
We use Pygments to color source code inside of documentation.
Unfortunately, Pygments requires Python 2, not Python 3, so if you don't
have this installed or this is not accessible, then you can't run YARD
locally to preview docs. The best way to get Python 2 is through
[`pyenv`][1]. `pyenv` will read from `.python-version` if it is set.
This commits adds that file.

[1]: https://github.com/pyenv/pyenv