Fix UI issues with generated docs
The current docsite isn't as usable as it should be: clicking on the nav
links should show a popover when in fact it goes to another page. This
is happening because some of the JavaScript is failing on the docsite
due to some old-school code (we're talking early 2000's DHTML stuff)
which was leftover from YARD's default layout which interacts with
frames. We don't use frames on the docsite so there is no need to have
this code.

Also this commit tweaks the CSS for the class list to remove extra
indentation from the bullet points.