Ensure allow_nil on inclusion returns self
It seems like we weren't testing allow_nil or allow_blank at all
properly. A curious thing popped up when I re-enabled these tests: it
also seems like for in_range, it doesn't matter if the validation has
allow_nil/allow_blank on it or not — adding allow_nil/allow_blank to the
inclusion matcher will work. Not sure if this is a regression or not.