Don't include setup script in global executables
All of the files in `bin/` were included in the list of executables in
the gemspec. This is Very Bad and means that `setup` was getting
installed as a global executable when a user installed the gem on their
computer.

To fix this, I've done an audit and removed all extraneous files from
the installed gem. It should now only include:

* `docs`
* `lib`
* `MIT-LICENSE`
* `README.md`
* `shoulda-matchers.gemspec`