Clear up confusion w/ presence around belongs_to associations
Now that `belongs_to` associations add a presence validation
automatically, if you have a test such as the following:

    class StudentBook < ApplicationRecord
      belongs_to :student
    end

    RSpec.describe StudentBook do
      it { is_expected.not_to be_valid }
      it { is_expected.to validate_presence_of(:student) }
    end

then the test for presence of :student will fail, because the validation
message on the automatic presence validation is different than the usual
message. The solution here, of course, is to just use `belong_to`, but
that is not very obvious. So this commit updates the presence matcher to
remind users about using `belong_to` in a case such as this.