validate_presence_of(:active_storage_field) throws exception
When trying to test validation of presence against a active storage relation, shoulda tries to set the attribute as empty string. This makes the association fail to be set and breaks the test.

I've introduced a block that checks if the field is an active storage relation and only allows `nil` to be set to the object.

Fixes #1223