Fix presence matcher against an AM record
When using the presence matcher against an ActiveModel record, check to
ensure that the model is using the Attributes API and has explicitly
declared the attribute to be a string before attempting to set the
attribute to an empty string. Otherwise, the attribute may be getting
treated as an object in some point in the validation process, and
therefore setting it to an empty string would create problems.