Provide way to force CI run
Usually we only run CI on master or on pull requests that are based on
master, but this adds a way to bypass this check for specific commits.
If you add "[ci run]" or "[run ci]" to a commit then this will force a
build on a branch, even if there is no PR for it.