Add missing test, add rails version conditional, refactor message
This adds a missing test for when there are no scope methods, but this
was not expected.

Additionally, it adds a refactoring for the message that is produced for
the errors.

Finally, it only runs the tests if the rails version is 6 or greater
(when this feature was made available).