Update references from master to main (#1476)
* Update references to `master` on documentation

The master branch has been renamed to main, but the docs
still reference it as master.

* Update the changelog URI

The changelog URI should be pointed to the main branch, where it is
updated. The master branch is archived and has not been updated since
version 5.0.0.

* Update the main branch on Github Actions config

Since shoulda-matchers 5.1.0, the main branch of the repository
is `main`.