Compatibility changes to `Submatchers` and `ComparisonMatcher`
We needed to make sure the ComparisonMatcher will be working for both
ValidateComparisonOfMatcher and ValidateNumericalityOfMatcher, some
changes were required to make it work properly.

We also changed the order of the submatchers, where the `DissalowMatcher`
of the `ValidateNumericalityOfMatcher` will always be evaluated first,
because of that change it was required some minor changes in the specs.