Improve `have_db_index` to better handle columns with multiple indexes
The `have_db_index` matcher currently matches on the first index
that has the matching column names. But when used with the `unique`
qualifier that can result in the "wrong" index being matched causing
the test to fail. For example, one unique BTree index, and another
non-unique partial or GIN index.