feat: Implement ValidateComparisonOfMatcher class
This commit implements the base of the validate_comparion_of_matcher.
Creating the `ValidateComparisonOfMatcher` class and specs for the
submatchers `is_greater_than`, `is_greater_than_or_equal_to` and
`is_less_than`. In the next commits we'll implement the specs for the
other submatchers and for combinations of submatchers.

We are also missing the specs and implementation for comparing with
other values than just pure integers. We need to implement an way to
compare with Procs and also symbols.