feat: Implement support for comparing between string values
In this commit we added the support for making validations using the
`validate_comparison_of` matcher for string values. Ruby compares
strings using their ASCII/Unicode value, so for example:

```ruby
'aaa' > 'bbb' # false
'bbb' > 'aaa' # true
```

In this example `'aaa'` gets converted to an array with the unicode
values for the letters
```ruby
'aaa' #=> [97, 97, 97]
```

Ruby compares two strings character by character based on their Unicode
code point values. When comparing two strings, it compares the
corresponding characters at each position until it finds a difference.

If the characters at the current position are the same, it moves on to
the next position and continues the comparison. If the characters at
the current position are different, it compares their Unicode code
point values to determine which character is greater. The string with
the greater character at the first differing position is considered
greater than the other string.