feat: Implement comparison of matcher (#1552)
* Extract comparison matcher from the NumericalityMatcher module

The comparison matcher will be used now on both
ValidateNumericalityOfMatcher and on the new
ValidateComparisonOfMatcher, that's why we removed it from the
NumericalityMatcher module.

* Compatibility changes to `Submatchers` and `ComparisonMatcher`

We needed to make sure the ComparisonMatcher will be working for both
ValidateComparisonOfMatcher and ValidateNumericalityOfMatcher, some
changes were required to make it work properly.

We also changed the order of the submatchers, where the `DissalowMatcher`
of the `ValidateNumericalityOfMatcher` will always be evaluated first,
because of that change it was required some minor changes in the specs.

* fix: Get failure_message_when_negated from DisalllowMatcher

This commit intends to instead of getting the first submatcher
failure_message_when_negated message, to get from the DisallowMatcher
first because the messages are more easy to read and to debug.

* feat: Implement ValidateComparisonOfMatcher class

This commit implements the base of the validate_comparion_of_matcher.
Creating the `ValidateComparisonOfMatcher` class and specs for the
submatchers `is_greater_than`, `is_greater_than_or_equal_to` and
`is_less_than`. In the next commits we'll implement the specs for the
other submatchers and for combinations of submatchers.

We are also missing the specs and implementation for comparing with
other values than just pure integers. We need to implement an way to
compare with Procs and also symbols.

* feat: Implement support for comparing between string values

In this commit we added the support for making validations using the
`validate_comparison_of` matcher for string values. Ruby compares
strings using their ASCII/Unicode value, so for example:

```ruby
'aaa' > 'bbb' # false
'bbb' > 'aaa' # true
```

In this example `'aaa'` gets converted to an array with the unicode
values for the letters
```ruby
'aaa' #=> [97, 97, 97]
```

Ruby compares two strings character by character based on their Unicode
code point values. When comparing two strings, it compares the
corresponding characters at each position until it finds a difference.

If the characters at the current position are the same, it moves on to
the next position and continues the comparison. If the characters at
the current position are different, it compares their Unicode code
point values to determine which character is greater. The string with
the greater character at the first differing position is considered
greater than the other string.

* feat: Implement support for comparing between date values

In this commit we added the support for making validations using the
`validate_comparison_of` matcher for date values.

* feat: Implement `allow_nil`, `strict` and `with_message` qualifiers

This commit inteds to implement the `allow_nil`, `strict` and
`with_message` qualifiers on top of ComparisonMatcher.

These qualifiers were copy/pasted from the NumericalityMatcher on
previous commits and some of them required some minor adjustments to
work properly.

* doc: Add documentation to `validate_comparison_of` matcher

* fix: Revert change for failure_message_when_negated