feat: Allow length validation on associations (#1569)
This commit allows the length matcher to be used on associations. It
does this by checking if the attribute is an association, and if so, it
uses the associations as the attribute to validate.

This commit also test for the length matcher on associations (has_many and has_many through).

I want to give credit to @prashantjois for the initial work on this
feature. I took his work and expanded on it.

Closes #1007