Move Elliot to alumni (#1577)
Hello everyone, as you are probably all too aware, I have not been very
active as a maintainer over the past few years. The reality is that in
my day job, I don't work with Ruby or Rails at all, and outside of work,
my time is spent less on side projects and more on family and other
sorts of commitments. Because of this, I feel like it is high time that
I officially step away from this project.

I want to thank thoughtbot for giving me the opportunity to maintain
this gem for so many years (I think I started in 2013!). I also want to
thank Pedro and Matheus for keeping the candle burning and continuing to
merge new changes, and of course those of you who still find this gem
useful in your work even now.

If you still need me for any reason, you can always ping me on Twitter
or Mastodon (links in my bio).

-- Elliot