fix: Adjust negative form of `ValidateNumericalityMatcher`
This commit adjusts the negative form of the `ValidateNumericalityMatcher`
so that is does not use the `does_not_match?` method, but actually uses the
negative result of the `matches?` method. This is easy to understand and
does not require any special handling for the negative form.

Previously this matcher on the negative form was always returning `true`,
as there were no `does_not_match?` method defined, which caused it to
use the `does_not_match?` method from the `ValidationMatcher` parent class,
which always returns `true`, this commit fixes that.