Support association strict_loading option
`strict_loading` was added to [Rails 6.1] to prevent lazy loading of
associations.

As adding it to an association declaration can have a massive impact on
the way the record and its association is treated, it can be useful to
ensure in a test suite the presence of this option.

This adds support for adding the `strict_loading` option to an
association declaration.

[Rails 6.1]: https://github.com/rails/rails/pull/37400

Co-authored-by: Jose Blanco @laicuRoot <jose.blanco@thoughtbot.com>