fix: Polymorphic + STI uniqueness check
Instead of just initializing a new class that inherits from the parent
class, we are initializing a new class but also overriding some
method definitions. That is necessary because the Rails guides
recommend that whenever you have a polymorphic association and
STI you should override the `..._type` column to use the type name
of the base STI class.